Android Switch Button Example
Android’s Switch Switch is a two-state toggle switch widget that can
select between two options.The user may drag the “thumb” back and forth
to choose the selected option, or simply tap to toggle as if it were a
checkbox. this very useful widget for toggle event like “ON or “OFF”. We
can set two event on switch button when goes to On or Off state.